Moved content.json to data folder
