+++ .eslintignore
@@ -1,2 +1,2 @@
 src/scripts
-src/Content.json
+src/data/content.json
