+++ src/components/About/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Content from '../../Content.json';
+import Content from '../../data/content.json';
 import ProfileImg from '../../assets/misc/profile.jpg';
 
 const About = ({ backgroundColor, panelColor }) => (
