+++ src/components/Education/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Content from '../../Content.json';
+import Content from '../../data/content.json';
 import McMasterImg from '../../assets/misc/mcmaster-logo.png';
 
 const EDUCATION = Content.education;
